+++ src/main/java/com/facebook/hive/metastore/client/HiveMetastore.java
@@ -26,4 +26,6 @@ public interface HiveMetastore extends Closeable, FacebookHiveMetastore
 {
     @Override
     void close();
+
+    boolean isConnected();
 }